ALEPH-1 moved test case to not use injection
