--- aleph2_access_manager/config.properties
@@ -1,10 +1,0 @@
-#access manager
-access_manager.AccessManager.service=com.ikanow.aleph2.access_manager.data_access.AccessContext
-
-#data service needs name, full path to overriding interface (if exists), full path to new service
-data_service.SecurityService.interface=com.ikanow.aleph2.data_model.interfaces.shared_services.ISecurityService
-data_service.SecurityService.service=com.ikanow.aleph2.security_service.SecurityService
-
-data_service.SampleCustomService.service=com.ikanow.aleph2.access_manager.data_access.SampleCustomService
-
-someproperty=somevalue
