+++ aleph2_access_manager/resources/application.conf
@@ -8,7 +8,7 @@ access_manager.service=com.ikanow.aleph2.access_manager.data_access.AccessContex
 #these data services can be accessed by calling AccessDriver.getAccessContext(); and requesting the service by the
 #name you give them
 #override the security service
-data_service.SecurityService.interface=com.ikanow.aleph2.data_model.interfaces.shared_services.ISecurityService
-data_service.SecurityService.service=com.ikanow.aleph2.security_service.SecurityService
+#data_service.SecurityService.interface=com.ikanow.aleph2.data_model.interfaces.shared_services.ISecurityService
+#data_service.SecurityService.service=com.ikanow.aleph2.security_service.SecurityService
 #add a custom service
 data_service.SampleCustomService.service=com.ikanow.aleph2.access_manager.data_access.SampleCustomService
