+++ aleph2_access_manager/test/com/ikanow/aleph2/access_manager/data_access/SampleCustomService.java
@@ -1,0 +1,13 @@
+package com.ikanow.aleph2.access_manager.data_access;
+
+import com.google.inject.Provides;
+import com.ikanow.aleph2.data_model.interfaces.data_layers.IDataService;
+
+public class SampleCustomService implements IDataService {
+
+//	@Provides
+//	public IDataService provideService() {
+//		return new SampleCustomService();
+//	}
+
+}
